+++ packages/client-lib/src/utils/api/index.ts
@@ -24,27 +24,52 @@ export class ApiService {
     return baseUrl;
   }
 
+  /**
+   * If the endpoint will not return any content then `Response` should be
+   * `null`.
+   *
+   * @example
+   *
+   * ```
+   * const data = apiService.request<null>({ ... })
+   * ```
+   */
   async request<Response>({
     method,
     endpoint,
     params,
     payload,
-  }: RequestParams) {
+  }: RequestParams): Promise<Response> {
+    const url = this.makeUrl(endpoint, params);
     const init = {
       method,
-      url: this.makeUrl(endpoint, params),
-      payload,
+      body: JSON.stringify(payload),
       headers: {
+        "Content-Type": "application/json",
         ...this.authHeader,
       },
     };
 
-    const response = await fetch(`${this.baseUrl}${endpoint}`, init);
-    const jsonData = await response.json();
+    const response = await fetch(url, init);
 
-    if (response.ok) return jsonData as Response;
+    const contentHeader = response.headers.get("Content-Length");
+    const hasContent = contentHeader && contentHeader !== "0";
 
-    throw new Error(JSON.stringify(jsonData));
+    // The caller should know whether or not the response has content.
+    // If it does, the `Response` type will be set as non-null.
+    // If it does not, the `Response` type will be set as `null`.
+    // Hence the value of responseJson will always match the caller's intention.
+    const responseJson: Response = hasContent ? await response.json() : null;
+
+    if (response.ok) return responseJson;
+
+    const error = {
+      status: response.status,
+      statusText: response.statusText,
+      body: responseJson,
+    };
+
+    throw new Error(JSON.stringify(error));
   }
 
   authenticate(userId: string): void {